"Testing of my_bookstore"

+++ src/main/java/mate/academy/spring/online/bookstore/service/book/BookServiceImpl.java
@@ -55,6 +55,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteBookById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't delete book by id " + id);
+        }
         bookRepository.deleteById(id);
     }
 
@@ -62,6 +65,16 @@ public class BookServiceImpl implements BookService {
     public BookDto updateBook(Long id, CreateBookRequestDto requestDto) {
         Book bookFetched = bookRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Book not found by id " + id));
+
+        List<Long> existingCategories = requestDto.getCategories().stream()
+                .filter(categoryRepository::existsById)
+                .toList();
+
+        if (existingCategories.isEmpty()) {
+            throw new EntityNotFoundException("Provided categories do not exist: "
+                    + requestDto.getCategories());
+        }
+
         bookMapper.updateBookFromDto(requestDto, bookFetched);
         return bookMapper.toDto(bookRepository.save(bookFetched));
     }
@@ -72,4 +85,5 @@ public class BookServiceImpl implements BookService {
         return bookRepository.findAll(bookSpecification, pageable)
                 .map(bookMapper::toDto);
     }
+
 }
